msg -> GetName() crashes, WTF. So does Persistent_GetName()!
Apparently there doesn't appear to be a direct way to safely validate message types, as any attempt to call GetName or Persistent_GetName on a message
crashes DromEd instantly. I need to investigate this and see if I can work out what in Eris' name is going on, in the meantime I've removed the checks
and just hope the user doesn't mess things up!